+++ crates/cli/src/commands/bin.rs
@@ -1,5 +1,5 @@
 use crate::tools::{create_tool, ToolType};
-use proto_core::{detect_version, Manifest};
+use proto_core::detect_version;
 use starbase::SystemResult;
 
 pub async fn bin(
@@ -8,14 +8,14 @@ pub async fn bin(
     use_shim: bool,
 ) -> SystemResult {
     let mut tool = create_tool(&tool_type).await?;
-    let manifest = Manifest::load(tool.get_manifest_path())?;
-    let version = detect_version(&tool, &manifest, forced_version).await?;
+    let manifest = tool.get_manifest()?;
+    let version = detect_version(&tool, manifest, forced_version).await?;
 
     tool.resolve_version(&version).await?;
     tool.find_bin_path().await?;
 
     if use_shim {
-        tool.create_shims().await?;
+        tool.setup_shims(true).await?;
 
         if let Some(shim_path) = tool.get_shim_path() {
             println!("{}", shim_path.to_string_lossy());