+++ crates/cli/src/commands/global.rs
@@ -1,12 +1,12 @@
 use crate::tools::{create_tool, ToolType};
-use proto_core::{color, Manifest};
+use proto_core::color;
 use starbase::SystemResult;
 use tracing::{debug, info};
 
 pub async fn global(tool_type: ToolType, version: String) -> SystemResult {
-    let tool = create_tool(&tool_type).await?;
+    let mut tool = create_tool(&tool_type).await?;
 
-    let mut manifest = Manifest::load(tool.get_manifest_path())?;
+    let manifest = tool.get_manifest_mut()?;
     manifest.default_version = Some(version.clone());
     manifest.save()?;