+++ crates/core/src/shimmer.rs
@@ -9,6 +9,8 @@ use tinytemplate::error::Error as TemplateError;
 use tinytemplate::TinyTemplate;
 use tracing::debug;
 
+pub const SHIM_VERSION: u8 = 1;
+
 #[derive(Serialize)]
 pub struct Context {
     bin_path: PathBuf,
@@ -179,11 +181,6 @@ impl ShimBuilder {
 
     fn do_create(&self, shim_path: PathBuf, contents: &str) -> Result<PathBuf, ProtoError> {
         let shim_exists = shim_path.exists();
-
-        // Temporary until we can version shims!
-        if shim_exists {
-            return Ok(shim_path);
-        }
 
         if let Some(parent) = shim_path.parent() {
             fs::create_dir_all(parent)?;